fixed bug where cards wouldn't invalidate between teams
