+++ package.json
@@ -85,7 +85,7 @@
     "babel-runtime": "6.18.0",
     "bcrypt": "^0.8.7",
     "body-parser": "1.15.2",
-    "cashay": "^0.21.0",
+    "cashay": "^0.21.1",
     "cheerio": "^0.22.0",
     "compression": "1.6.2",
     "cors": "2.8.1",
