+++ src/universal/components/ProjectColumns/ProjectColumns.js
@@ -4,9 +4,8 @@ import {css} from 'aphrodite-local-styles/no-important';
 import ui from 'universal/styles/ui';
 import {columnArray} from 'universal/utils/constants';
 import ProjectColumn from 'universal/modules/teamDashboard/components/ProjectColumn/ProjectColumn';
-import PrivateDragState from 'universal/utils/PrivateDragState';
+import privateDragState from 'universal/dnd/PrivateDragState';
 
-const privateDragState = new PrivateDragState();
 const ProjectColumns = (props) => {
   // myTeamMemberId is undefined if this is coming from USER_DASH
   // TODO we only need userId, we can compute myTeamMemberId
